pyproject.toml
[project]
name = "watermarking-contrastive-learning"
version = "0.1.0"
description = ""
authors = [
    {name = "Admir"}
]
readme = "README.md"
requires-python = "3.10.12"
dependencies = [
    "torch (>=2.8.0,<3.0.0)",
    "torchvision (>=0.23.0,<0.24.0)",
    "torchaudio (>=2.8.0,<3.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "mutagen (>=1.47.0,<2.0.0)",
    "torchcodec (>=0.6.0,<0.7.0)",
    "pystoi (>=0.4.1,<0.5.0)",
    "librosa (>=0.11.0,<0.12.0)",
    "pesq (>=0.0.4,<0.0.5)",
    "black (>=25.1.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

